create discovery.rs
